app/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
python-json-logger==2.0.7
prometheus-flask-exporter==0.22.4
# OpenTelemetry Dependencies
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp